requirements.txt
SQLAlchemy>=2.0
alembic>=1.12
requests
python-dotenv
pytest
